정오와 자정을 넘는 예약일때 버그 수정

+++ app/Http/Controllers/ReservationController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Reservation;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 use Rap2hpoutre\FastExcel\FastExcel;
 
 class ReservationController extends Controller
@@ -41,19 +42,21 @@ class ReservationController extends Controller
 
         // insert
         foreach ($rows as $row) {
-            $insertData = $this->transform($row);
-            if ($row['상태'] == '확정' && $row['결제상태'] == '결제완료') {
-                try {
-                    Reservation::create($insertData);
-                } catch (\PDOException $e) {
+            if($row['상태']) {
+                $insertData = $this->transform($row);
+                if ($row['상태'] == '확정' && $row['결제상태'] == '결제완료') {
+                    try {
+                        Reservation::create($insertData);
+                    } catch (\PDOException $e) {
 
+                    }
                 }
-            }
 
-            if ($row['상태'] == '취소') {
-                $reservation = Reservation::where($insertData)->first();
-                if ($reservation) {
-                    $reservation->delete();
+                if ($row['상태'] == '취소') {
+                    $reservation = Reservation::where($insertData)->first();
+                    if ($reservation) {
+                        $reservation->delete();
+                    }
                 }
             }
         }
@@ -112,7 +115,7 @@ class ReservationController extends Controller
         $result['name'] = $row['예약자'];
 
         $result['phone'] = mb_substr($row['전화번호'], - 4, 4);
-        $date = mb_substr($row['이용일시'], 0, strpos($row['이용일시'], '('));
+        $date = $fromDate = $toDate = mb_substr($row['이용일시'], 0, strpos($row['이용일시'], '('));
         $time = mb_substr($row['이용일시'], strpos($row['이용일시'], ')'));
 
         $timeArray = explode(' ', $time);
@@ -122,13 +125,24 @@ class ReservationController extends Controller
         if ($timeArray[0] == '오후') {
             $fromArray[0] = $fromArray[0] + 12;
             $toArray[0] = $toArray[0] + 12;
+
+            if($toArray[0] < $fromArray[0]) {
+                $toArray[0] = $toArray[0] - 12;
+                $toDateCarbon = Carbon::createFromFormat('y. m. d.', $toDate);
+                $toDate = $toDateCarbon->format('y. m. d.');
+            }
+        } else {
+            if ($toArray[0] < $fromArray[0]) {
+                $toArray[0] = $toArray[0] + 12;
+            }
         }
+
         $fromString = implode(':', $fromArray);
         $toString = implode(':', $toArray);
 
         //dd($date);
-        $result['from'] = Carbon::createFromFormat('y. m. d. H:i', $date . $fromString);
-        $result['to'] = Carbon::createFromFormat('y. m. d. H:i', $date . $toString);
+        $result['from'] = Carbon::createFromFormat('y. m. d. H:i', $fromDate . $fromString);
+        $result['to'] = Carbon::createFromFormat('y. m. d. H:i', $toDate . $toString);
 
         return $result;
     }
